atualizando o projeto

+++ app.php
@@ -12,7 +12,7 @@
     ini_set('memory_limit', '-1');
     ini_set('max_execution_time', 300); //300 seconds = 5 minutes
     ini_set("upload_max_filesize", "1024M");
-    // date_default_timezone_set("Brazil/East"); //Definindo timezone padrão
+    date_default_timezone_set("America/Manaus"); //Definindo timezone padrão
 
     $_POST['metodo']($raiz);
 
@@ -27,9 +27,21 @@
 
         // comparando as senhas
         if ($data['usuario'] === $user && $data['senha'] === $pass) {
+            // salvando o log
+            $resp = updateLog("Acesso ao sistema → permitido.");
+            if ($resp['success']==false) die(json_encode($resp));
+
             $resp['success'] = true;
             $resp['msg'] = "autenticado";
             $resp['data'] = array("usuario" => $data['usuario'], "autenticado" => true, "path" => $raiz);
+        }else{
+            // salvando o log
+            $resp = updateLog("Acesso ao sistema → recusado.");
+            if ($resp['success']==false) die(json_encode($resp));
+
+            $resp['success'] = false;
+            $resp['msg'] = "Acesso Negado, Usuário ou senha inválidos!";
+            die(json_encode($resp));
         }
 
         echo json_encode($resp);
@@ -101,7 +113,11 @@
                         $resp["msg"] = "Ocorreu um erro ao tentar enviar";
                         $resp["success"] = false;
                         die (json_encode($resp));
-                    } 
+                    }
+
+                    // salvando o log
+                    $resp = updateLog("Arquivo '$target_file' → enviado.");
+                    if ($resp['success']==false) die(json_encode($resp));
                 }
             }
 
@@ -116,16 +132,16 @@
         echo json_encode($resp);
     }
 
-    function downloadArquivo () {
-        $resp = array("success"=>false, "msg"=>"", "data"=>"");
-    }
-
     function deletar () {
 
         $resp = array("success"=>false, "msg"=>"", "data"=>"");
 
         $path = $_POST['path'];
         unlink($path);
+
+        // salvando o log
+        $resp = updateLog("Arquivo '$path' → removido.");
+        if ($resp['success']==false) die(json_encode($resp));
 
         $resp['success'] = true;
 
@@ -155,6 +171,10 @@
             die(json_encode($resp));
         }
 
+        // salvando o log
+        $resp = updateLog("Diretório '$path' → criado.");
+        if ($resp['success']==false) die(json_encode($resp));
+
         $resp['success'] = true;
         $resp['msg'] = "Diretório criado com sucesso.";
 
@@ -176,8 +196,19 @@
             function _delTree($dir) {
                 $files = array_diff(scandir($dir), array('.','..'));
                 foreach ($files as $file) {
-                (is_dir("$dir/$file")) ? _delTree("$dir/$file") : unlink("$dir/$file");
+                    if (is_dir("$dir/$file")) {
+                         _delTree("$dir/$file");
+                    }else{
+                       unlink("$dir/$file");
+                        // salvando o log
+                        $resp = updateLog("Arquivo '$dir/$file' → removido.");
+                        if ($resp['success']==false) die(json_encode($resp));
+                    }
                 }
+                // salvando o log
+                $resp = updateLog("Diretório '$dir' → removido.");
+                if ($resp['success']==false) die(json_encode($resp));
+
                 return rmdir($dir);
             }
             _delTree($path);
@@ -190,6 +221,61 @@
 
         $resp['success'] = true;
         $resp['msg'] = "Diretório criado com sucesso.";
+
+        echo json_encode($resp);
+    }
+
+    function updateLog ($log) {
+
+        $resp = array("success" => false, "msg" => "");
+
+        try {
+            // criando/escrevendo arquivo de log
+            $fp = fopen("log.txt", "a+");
+            $d = new DateTime();
+            fwrite($fp, $d->format("d/m/Y H:i:s") . " - $log\r\n");
+            fclose($fp);
+        }catch (Exception $e) {
+            $resp['success'] = false;
+            $resp['msg'] = "Ocorreu um erro ao salvar o log: " . $e;
+            die(json_encode($resp));
+        }
+
+        $resp['success'] = true;
+        $resp['data'] = true;
+        $resp['msg'] = "O log foi realizado com sucesso.";
+
+        return $resp;
+    }
+
+    function getLog ($raiz) {
+
+        $resp = array("success" => false, "msg" => "");
+
+        try {
+            // verificando se o arquivo de log existe
+            if (!file_exists("log.txt")) {
+                $resp['success'] = true;
+                $resp['data'] = array();
+                $resp['msg'] = "O log não existe.";
+                die(json_encode($resp));
+            }
+
+            // lendo log
+            $conteudo = file("log.txt");
+            for ($i=0; $i<count($conteudo); $i++) {
+                $conteudo[$i] = str_replace("$raiz/", "raiz/", $conteudo[$i]);
+            }
+
+        }catch (Exception $e) {
+            $resp['success'] = false;
+            $resp['msg'] = "Ocorreu um erro ao salvar o log: " . $e;
+            die(json_encode($resp));
+        }
+
+        $resp['success'] = true;
+        $resp['data'] = array_reverse($conteudo);
+        $resp['msg'] = "O log foi carregado com sucesso.";
 
         echo json_encode($resp);
     }
